+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 5.65.0 - 2020-07-06
 ### Added
 - `valentineScript` to nice servant and CE.
+- `notTrait` to svt search.
 
 ## 5.64.0 - 2020-06-17
 ### Added